Add Cloud Linux Software to copyright notices and license
Custody of this library transferred to CLS with an IP
acquisition in 2023 from Patchman B.V., and this
change reflects that stewardship handover.

In 2025, a copy of the repository was provided to Jelmer
Verkleij with the express purpose of publishing it
publicly and open-source. This commit is preparation
for that move to include attribution where appropriate
for its past stewards. However, since this process
happens outside of CLS and is done in a private capacity
unconnected to CLS, there is no CLS-bound URL to update
the metadata with.